refactor: util 을 service 로 변경 (#287)
