+++ src/main/java/com/smunity/server/global/scheduler/StatSendScheduler.java
@@ -1,8 +1,8 @@
 package com.smunity.server.global.scheduler;
 
 import com.smunity.server.global.common.dto.StatResponseDto;
+import com.smunity.server.global.common.service.SlackService;
 import com.smunity.server.global.common.service.StatService;
-import com.smunity.server.global.common.util.SlackUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -14,12 +14,12 @@ import static com.smunity.server.global.common.logging.Loggers.event;
 public class StatSendScheduler {
 
     private final StatService statService;
-    private final SlackUtil slackUtil;
+    private final SlackService slackService;
 
     @Scheduled(cron = "0 0 0 * * *")
     public void sendStatistics() {
         StatResponseDto responseDto = statService.getStatistics();
-        slackUtil.sendMessage(responseDto);
+        slackService.sendMessage(responseDto);
         event.info("[StatSendScheduler] event=sendStatistics status=success payload={}", responseDto);
     }
 }
